Move reducers into src
