+++ src/reducers/loading.js
